Upgrade Grails 3.2.7 --> 3.2.8
Of course, Grails 3.2.8 comes out the very next day after I upgraded to
Grails 3.2.7.

I compared two blank projects and applied the differences to the
matching files in this project.

    mkdir grails3.2.7
    sdk use grails 3.2.7
    (cd grails3.2.7; grails create-app blank)

    mkdir grails3.2.8
    sdk use grails 3.2.7
    (cd grails3.2.8; grails create-app blank)

    diff -r grails3.2.7/blank grails3.2.8/blank

My version of `application.yml` was a little bit behind, but it was
mostly sections moving around.  Nothing of consequence.