Thoughts on the app as a set of microservices
Wrote an overview diagram using Google Docs and exported it as
a `.png` file.

Wrote some sequence diagrams using `sdedit` but couldn't export
them as graphics file.  Could be a Java version mismatch between
the Java 8 I'm using versus the Java 6 it is expecting.

I will have to figure out how to reference the diagrams from the
GitHub Pages for the project.