Update and fix Passport integration
Somehow, encoding passwords was never returning.  I had to do a lot
of back and forth to get it working again.

Changed the password encoding algorithm to stronger encoding, and
save them as Base-64-encoded instead of hexadecimal.

Added a `/api/resetPassword` endpoint so I can overwrite the bogus
`salt` and `hash` that I hardcoded in the user generation script.

Brought the style closer to ES6, such as using arrow functions and
moving away from `var` variables.

Changed `password` to `hash` in the database schema.

         1111111111222222222233333333334444444444555555555566666666667
1234567890123456789012345678901234567890123456789012345678901234567890