Remove reviews list fields from User and Book models
I was managing the association in `Review` and in `User` and `Book`.
That was redundant.  I removed all review management from `User`
and from `Book`.  This way, when I created or delete a `Review`,
it is implicitly added or removed from the subordinate lists.

I also renamed the model files.  They were the only ones with
plural names for the model classes.