Fix broken tests
Normally, the test would create a JWT and pass it along as a request
header.  But the current version of RSpec does not handle the headers
passed as part of the request, so we have to hack the controller to
make it believe it received the appropriate `Authorization` header with
the JWT in it.