Use SQL variables instead of global Perl variables in DB setup
I didn't know how to retrieve the ID of inserted rows in MySQL, so I had
to set IDs explicitly and track their value in the Perl script that
generates the DB setup script.

While setting up a new MySQL installation, I looked at the
`mysql_tzinfo_to_sql` command that sets up the timezone tables.  This
line drew my attention:

    SET @time_zone_id= LAST_INSERT_ID();

It showed how to get the ID from the previous `INSERT` and it showed how
to save it for future use.  The following lines showed how to use
`@time_zone_id` in subsequent SQL statements.

Now, there are no hard-coded IDs in the DB setup script for MySQL, just
like the ones for MongoDB in other branches.

I was also able to learn how to make the DB script report its progress
as it runs.