Connect to MySQL using TCP instead of local socket
In my initial development environment, I had MySQL installed locally,
so the `mysql2` gem was able to use a local socket to talk to it.  In my
alternate development environment, I use Docker to run MySQL.  Since the
database is now in a container, the app is no longer able to connect to
it using a socket.  But the container does open its TCP port, so the app
can connect to it through that.