Refactor uses of BookType's batch loaders
Use the cached book record instead of the underlying object.  They are
the same, but this exemplifies a different way to use batch loading.
Also, use slightly more compact notation to extract strings from helper
models for a book's authtors and year(s) of publication.  I find it
somewhat less elegant, but it may be more Ruby-ish.  And, it lines up
with the Java Stream API.  Not quite Groovy's `*.` operator, though.