Name all ID fields "id"
Books had `bookId`, reviews had `reviewId`, users had `userId`.  It
helped recognize IDs as they were flying around, but it was not
consistent with other implementations.

One interesting bit is how it simplified search indexing.  The search
index would store structures with `__typename`, `bookId`, `reviewId`,
and `userId` and return them as search results.  The gateway would use
the `__typename` and only use the relevant _*Id_ field.  But now, the
structure is simply a `__typename` and an `id`.  No matter the type of
result, it's ID field is always `id`.