Rename association descriptor
The name `association_name` hints at a single, direct association.  The
call to `Preloader#preload` supports a much richer format for
designating multiple associations and sub-associations at once.  We
change the name of the parameter to `associations` to draw a direct link
between this parameter and what gets passed to `Preloader#preload`.