Use stand-alone Neo4J
The embedded Neo4J won't start.  It complains that:

```
Embedded Neo4j server was configured but 'neo4j-harness' classes not found on classpath.
```

I tried to include a dependency on `neo4j-harness`, but Gradle is unable
to resolve it.

I replaced the embedded Neo4J config with the credentials to talk to a
local Neo4J instance (running in Docker).  Now, the server hangs during
startup.