Add simple entity and controller
I had to downgrade Neo4j from `4.2` to `3.5` because the ORM is using a
CYPHER syntax that is no longer valid.  It was deprecated in `3.0` and
removed in `4.0`.  I don't quite get why the drivers have not been
updated.

I can `GET /users` when there are no `User` nodes, but the moment I
create one, it starts to `500` with this error message coming from the
driver:

```
Cannot access records on this result any more as the result has already
been consumed or the query runner where the result is created has
already been closed.
```

I can `POST /users` and `GET /users/<id>` and `DELETE /users/<id>` just
fine.