Add a relationship modeled as a GORM entity
The documentation says to put it as a `hasMany` on the node entities,
but GORM is unable to load them from the nodes.  It complains that the
connection to the database has already been closed by the time it tries
to read data from it.

The `GET` action on `ReviewController` issues the correct CYPHER query,
but it is not receiving the data, for some reason.