Do not delete review before all validations
Prior to this change, we deleted the review with single `Review
.findOneAndDelete()` call, and *then* checked that the user was actually
authorized to delete that review.  This meant that if one user tried to
delete another user's review, they would get an error message but would
still have deleted that review.

Now, we make two calls to the database.  One `Review.findOne()` to load
the review so we can validate that the user is allowed to delete it, and
a separate call to `Review.deleteOne()` if they are allowed.