Refactor Review into a node in the graph
Invert the `:REVIEW` relationship into `:REVIEWER` and `:BOOK`
respectively.

Replace `hasMany` in `Book` and `User` with `hasOne` in `Review`.  Use
explicit getter in `Book` and `User` to fetch reviews.  The Jackson
renderer has issues with infinite recursion when it tries to convert
entities to JSON.  It's fine with a user that has no reviews, but
otherwise fails.

Many of the controllers have issues with reading data from the database
after the connection or transaction has completed.  I fixed it when
going from `Book` to `Title` by loading titles eagerly, but I don't want
to do that with all models and associations.

Use better HTTP response codes, where relevant.