Docker Compose is now in the Docker CLI
Use `docker compose` instead of `docker-compose`.  This new Docker
Compose does not pick up the override file automaticaly, so I had to
amend the commands to list the Docker Compose files explicitly.

The `restart.sh` script sometimes (i.e., most of the time) does not stop
listening to the logs for the "Server ready at" message.  For some
reason, the `break` command does not move processing to the next part of
the script.