Create search indices from Kafka messages
The `search` microservice listens for messages from the other services
and builds its indices accordingly.

I removed the `addIndex` function from `seed.sh`.  It was just a
placeholder until I could get the `search` microservice to be smart
enough to process entities directly.

I'm annoyed that I need to use a `switch` statement on the message type
in the single message handler for all messages across all topics.  This
is not very modular.