Order search results by weight
When searching for multiple words, the length of the matching words give
weight to the result and promote it towards the front of the result set.

The `search` microservice still doesn't take into account how many times
a word is referenced in the match.  Both "this sentence is blah" and
"blah blah blah" score the same for the word _blah_, that is 4.