Index words based on score
Each word in a corpus gets a score that represents how much of that
corpus the given word makes up.  That is, what proportion of the corpus
is made up of a given word.  The word _blah_ in the corpus "Blah" has a
score of `1.0` (the characters of _blah_ cover 4 out of 4 characters of
"Blah").  The same word _blah_ in the corpus "Blah blah blah" has a
score of about `0.85714` (the characters of _blah_ cover 12 out of 14
characters of the corpus).

When searching for multiple words, the score of the matching words give
weight to the result and promote it towards the front of the result set.