Remove DELETE CASCADE in the database
Let Rails deal with associations, so it can call the appropriate
callbacks when deleting entities.  It is less efficient than handling
associations in the database with a single `DELETE` statement, but it is
better from a business logic perspective.

I hadn't regenerated `schema.rb` in a while and newer versions of Rails
structure it slightly differently.  The relevant changes are on lines 73
and 74.