Adjust message intake for reviews service
The instances of the `reviews` microservice all need to receive each and
every `reviewAdded`, `reviewUpdated`, and `reviewRemoved` messages so
they can synchronized their internal state.  However, we don't need
every one of them to handle `bookRemoved` and `userRemoved` messages.
That would result in a flurry of `reviewRemoved` messages, all for the
same few reviews.  It is better that instances of the `reviews`
microservice use a consumer group so only one instance at a time needs
to process `bookRemoved` or `userRemoved` messages.