Add jwts microservice
This new microservice computes a JWT for a user.  I stil need to figure
out how to login a user (and add roles for authorizations).

In  "Production Ready GraphQL", author Marc-André Giroux recomends
against doing login in GraphQL and use a separate endpoint instead.  I
don't want to expose my `users` microservice and it's the one with all
the user-related information.  So, I will have to do it in GraphQL after
all.