Use JWT bearer token
The gateway passes the `Authorization` header with the bearer token, if
present, to the services.  The services validate the token and expose
its payload as part of the GraphQL context.

The `users` microservice adds a new `me` query that uses the JWT payload
to figure out who the current user is.

The `search` microservice includes the current user in the data it logs
to the `book-reviews.searches` topic.