Refactor login
Use a mutation on the `jwts` microservice.  The service needs to compile
and maintain a list of users with their credentials, so it listens to
the `book-reviews.users` topic to do it.

Added a `password` to the user structure.  The password is never
surfaced by the GraphQL schema.  It is not encrypted, though.