Remove top-level .ruby-version file
In the spirit of having microservices be as independent from each
other as possible, each microservice should maintain its own
`.ruby-version` file.  There is no usasge of Ruby at the project's
top-level, so there is no need for a `.ruby-version` file there.

It is a little repetitive to have these (identical) files in each
service folder, but I decided to keep them in order to maintain
each service's self-reliance.