Update dependencies
    npm update

It upgraded `mongoose` from `7.5.2` to `8.0.0`.  It brings in `bson` at
version `6.2`.  It also upgraded `mongodb-memory-server`, which I use
for testing, from `8.15.1` to `9.0.1`, which depends (indirectly) on
`bson` `5.5`.  This breaks some tests where `mongoose` is expecting a
`bson` `6.x.x` object, but it is instead getting a `bson` `5.x.x` object
from `mongodb-memory-server`.  I will leave the test in its broken state
for a little while and hope that `mongodb-memory-server` will upgrade to
`bson` `6.x.x` soon.