Silence controller callback problem
I'm seeing the following error when I try to access `/books/[:id]/reviews` or
`/users/[:id]/reviews`.  It doesn't complain for book authors or other parts,
only for reviews.

```
Started GET "/books/164/reviews" for ::1 at 2024-11-10 11:29:09 -0500
Processing by BookReviewsController#index as */*
  Parameters: {"book_id"=>"164"}
Completed 404 Not Found in 0ms (ActiveRecord: 0.0ms (0 queries, 0 cached) | GC: 0.0ms)

AbstractController::ActionNotFound (The create action could not be found for the :authenticate_user!
callback on BookReviewsController, but it is listed in the controller's
:only option.

Raising for missing callback actions is a new default in Rails 7.1, if you'd
like to turn this off you can delete the option from the environment configurations
or set `config.action_controller.raise_on_missing_callback_actions` to `false`.
):
```